+++ src/main/java/kr/hs/entrydsm/raisepercent/domain/teacher/code/exception/CodeNotFoundException.java
@@ -1,4 +1,4 @@
-package kr.hs.entrydsm.raisepercent.domain.code.exception;
+package kr.hs.entrydsm.raisepercent.domain.teacher.code.exception;
 
 import kr.hs.entrydsm.raisepercent.global.error.exception.ErrorCode;
 import kr.hs.entrydsm.raisepercent.global.error.exception.RaiseException;
