+++ src/main/java/kr/hs/entrydsm/raisepercent/domain/teacher/code/service/CodeReissueService.java
@@ -1,8 +1,8 @@
-package kr.hs.entrydsm.raisepercent.domain.code.service;
+package kr.hs.entrydsm.raisepercent.domain.teacher.code.service;
 
-import kr.hs.entrydsm.raisepercent.domain.code.domain.Code;
-import kr.hs.entrydsm.raisepercent.domain.code.domain.repositories.CodeRepository;
-import kr.hs.entrydsm.raisepercent.domain.code.facade.CodeFacade;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.domain.Code;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.domain.repositories.CodeRepository;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.facade.CodeFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
