+++ src/main/java/kr/hs/entrydsm/raisepercent/domain/teacher/code/service/CodeVerificationService.java
@@ -1,13 +1,13 @@
-package kr.hs.entrydsm.raisepercent.domain.code.service;
+package kr.hs.entrydsm.raisepercent.domain.teacher.code.service;
 
-import kr.hs.entrydsm.raisepercent.domain.code.domain.Code;
-import kr.hs.entrydsm.raisepercent.domain.code.domain.repositories.CodeRepository;
-import kr.hs.entrydsm.raisepercent.domain.code.exception.CodeNotFoundException;
-import kr.hs.entrydsm.raisepercent.domain.code.exception.CodeNotMatchException;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.domain.Code;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.domain.repositories.CodeRepository;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.exception.CodeNotFoundException;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.exception.CodeNotMatchException;
 import kr.hs.entrydsm.raisepercent.domain.teacher.domain.Teacher;
 import kr.hs.entrydsm.raisepercent.domain.teacher.domain.types.Role;
 import kr.hs.entrydsm.raisepercent.domain.teacher.facade.TeacherFacade;
-import kr.hs.entrydsm.raisepercent.domain.teacher.presentation.dto.request.VerifyTeacherRequest;
+import kr.hs.entrydsm.raisepercent.domain.teacher.code.presentation.dto.request.VerifyTeacherRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
