+++ src/main/java/kr/hs/entrydsm/raisepercent/domain/teacher/presentation/TeacherController.java
@@ -1,27 +1,17 @@
 package kr.hs.entrydsm.raisepercent.domain.teacher.presentation;
 
-import kr.hs.entrydsm.raisepercent.domain.code.service.CodeVerificationService;
-import kr.hs.entrydsm.raisepercent.domain.code.service.QueryCodeService;
-import kr.hs.entrydsm.raisepercent.domain.code.service.CodeReissueService;
-import kr.hs.entrydsm.raisepercent.domain.teacher.presentation.dto.request.VerifyTeacherRequest;
 import kr.hs.entrydsm.raisepercent.domain.teacher.service.GoogleAuthService;
 import kr.hs.entrydsm.raisepercent.domain.teacher.service.QueryGoogleAuthLinkService;
 import kr.hs.entrydsm.raisepercent.domain.user.presentation.dto.request.CodeRequest;
 import kr.hs.entrydsm.raisepercent.global.security.jwt.dto.TokenResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
 
 @RequiredArgsConstructor
 @RequestMapping("/teachers")
 @RestController
 public class TeacherController {
 
-    private final QueryCodeService queryCodeService;
-    private final CodeReissueService codeReissueService;
-    private final CodeVerificationService codeVerificationService;
     private final GoogleAuthService googleAuthService;
     private final QueryGoogleAuthLinkService queryGoogleAuthLinkService;
 
@@ -33,22 +23,6 @@ public class TeacherController {
     @PostMapping("/auth")
     public TokenResponse googleAuthLogin(@RequestBody CodeRequest codeRequest) {
         return googleAuthService.execute(codeRequest);
-    }
-
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PutMapping("/auth")
-    public void verifyTeacher(@RequestBody @Valid VerifyTeacherRequest request) {
-        codeVerificationService.execute(request);
-    }
-
-    @GetMapping("/code")
-    public String codeIssue() {
-        return queryCodeService.execute();
-    }
-
-    @PutMapping("/reissue")
-    public String codeReissue() {
-        return codeReissueService.execute();
     }
 
 }
